feat: 기본 조회 기간 설정
